requirements.txt
# Local LLM + RAG
ollama>=0.4.0
llama-index>=0.12.0
llama-index-llms-ollama>=0.5.0
llama-index-embeddings-ollama>=0.5.0
llama-index-vector-stores-chroma>=0.4.0

# Vector store
chromadb>=0.6.0

# Document parsing
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
python-pptx>=1.0.0
beautifulsoup4>=4.12.0
markdown>=3.7
chardet>=5.2.0

# OCR (requires: brew install tesseract poppler)
pytesseract>=0.3.10
pdf2image>=1.17.0

# Outlook 365
msal>=1.31.0
requests>=2.32.0

# UI
streamlit>=1.41.0

# Utilities
python-dotenv>=1.0.0
watchdog>=6.0.0
tqdm>=4.67.0
